Finished exercises 4.15 - 4.23

+++ Part4/blog_list/backend/controllers/blogs.js
@@ -1,43 +1,61 @@
 const blogsRouter = require('express').Router()
+const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
+const User = require('../models/user')
+
+
+
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
-  response.json(blogs)
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
+  response.json(blogs.map(blog => blog.toJSON()))
 })
 
 blogsRouter.post('/', async (request, response) => {
-  try {
-    const body = request.body
+  const body = request.body
 
-    const blog = new Blog({
-      title: body.title,
-      author: body.author,
-      url: body.url,
-      likes: body.likes
-    })
-    
-    const savedBlog = await blog.save()
-    response.status(201).json(savedBlog)
-  } catch (error) {
-    if (error.name === 'ValidationError') {
-      return response.status(400).json({ error: error.message })
-    }
-    response.status(500).json({ error: 'Something went wrong' })
+  const user = request.user
+
+  if (!request.user) {
+    return response.status(401).json({ error: 'user not found' })
   }
+
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes,
+    user: user
+  })
+   // Save the blog to the database and associate it with the user
+  const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+  response.status(201).json(savedBlog.toJSON())
 })
 
 blogsRouter.delete('/:id', async(request, response) => {
   const id = request.params.id
-  try {
-    const result = await Blog.findByIdAndDelete(id)
-    if (result) {
-      response.status(204).end()
-    } else {
-      response.status(404).json({ error: 'Blog not found' })
-    }
-  } catch (error) {
-    response.status(500).json({ error: 'Something went wrong' })
+  const user = request.user
+
+  const blog = await Blog.findById(id)
+  if (!blog) {
+    return response.status(404).json({ error: 'Blog not found' })
+  }
+
+   if (!request.user) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  if (blog.user.toString() !== user.id.toString()) {
+    return response.status(403).json({ error: 'only the creator can delete this blog' })
+  }
+
+  const result = await Blog.findByIdAndDelete(id)
+  if (result) {
+    response.status(204).end()
+  } else {
+    response.status(404).json({ error: 'Blog not found' })
   }
 })
 
@@ -50,15 +68,11 @@ blogsRouter.put('/:id', async (request, response) => {
     url: body.url,
     likes: body.likes
   }
-  try {
-    const updatedBlog = await Blog.findByIdAndUpdate(id, blog, { new: true })
-    if (updatedBlog) {
-      response.json(updatedBlog)
-    } else {
-      response.status(404).json({ error: 'Blog not found' })
-    }
-  } catch (error) {
-    response.status(500).json({ error: 'Something went wrong' })
+  const updatedBlog = await Blog.findByIdAndUpdate(id, blog, { new: true, runValidators: true })
+  if (updatedBlog) {
+    response.json(updatedBlog)
+  } else {
+    response.status(404).json({ error: 'Blog not found' })
   }
 })
